+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-04-24 18:35:44
-// Commit: refactor(core): decouple repository interface from concrete core service
+// Updated at: 2026-04-25 22:03:07
+// Commit: fix(core): fix memory leak by cancelling core subscription
 
 class coreComponentState {
-  final String commitRef = '643';
-  final String lastMessage = 'refactor(core): decouple repository interface from concrete core service';
-  final DateTime timestamp = DateTime.parse('2026-04-24T18:35:44+05:00');
+  final String commitRef = '692';
+  final String lastMessage = 'fix(core): fix memory leak by cancelling core subscription';
+  final DateTime timestamp = DateTime.parse('2026-04-25T22:03:07+05:00');
 }